+++ NEWS.md
@@ -1,3 +1,15 @@
+# erify 0.3.0
+
+* Add `is_n()`.
+* Add `check_contents()`.
+* Add `throw()`.
+* Deprecate `print_string()`.
+* Deprecate `Statement()` and `trigger()`, please use `throw()`.
+* Remove dependencies on package magrittr and rlang.
+* Add options `erify.bullets` and `erify.n`.
+* Require R >= 4.1.0.
+
+
 # erify 0.2.0
 
 ## Features